Guess annotation value types to improve aggregate queries over span data
A shortcoming of the Zipkin API is that it expects binary annotation values to
be strings ("binary annotations" correspond to "tags" in the OpenTracing spec).
So if you write something like

```
span := opentracing.StartSpanFromContext(ctx, "mySpan")
// ...
span.SetTag("num_segments", 22)
```

then client libraries will turn the integer `22` into the string `"22"`
when publishing the span.

Zipin's API decision here may make sense in a world oriented around single
traces, but is unfortunate if you'd like to do any sort of
numerical aggregate query over your span data (e.g., `MAX(num_segments)`...).

So in the proxy, try to guess in the proxy if binary annotation values
are actually bools, ints or floats, and convert them if so. I'm far from
thrilled with this, but don't really see a better way.