Allow spans to be routed to different datasets
Looking for feedback on this approach. It should be possible to
send different spans to different datasets from within the same
codebase. (In particular, it makes sense to send write traces and query
traces to different datasets in our internal deployment.)

In this patch, we do that by letting clients set a special tag, `honeycomb.dataset`, on
spans, and using that tag's value to override whatever default dataset
is configured in the opentracing proxy. There's some precedent for using
span tags to influence tracer behavior (see e.g. the `sampling.priority`
tag in
https://github.com/opentracing/specification/blob/master/semantic_conventions.md#span-tags-table.)

This does require more effort for clients, and I don't think it'll
be relevant to the majority of users. But we need it internally, so
putting this patch up in case anyone has opinions.