Create plain binaries as build artifacts
Update Travis CI to publish plain binary artifacts to the
`honeycomb-builds` S3 bucket, same as we do for honeytail, etc. Also
change how we publish docker images to be simpler and more consistent
with the rest of our integrations.

Not publishing deb/rpm packages yet, but can do that in a subsequent
patch if needed.